+++ src/main/java/med/voll/api/controller/PacienteController.java
@@ -4,8 +4,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import med.voll.api.paciente.DadosCadastroPaciente;
 
-import med.voll.api.medico.DadosCadastroPaciente;
 
 @RestController
 @RequestMapping("/pacientes")
