+++ src/main/java/med/voll/api/paciente/DadosCadastroPaciente.java
@@ -1,4 +1,4 @@
-package med.voll.api.medico;
+package med.voll.api.paciente;
 
 import med.voll.api.endereco.DadosEndereco;
